Time: 3 ms (84.85%) | Memory: 16.6 MB (12.86%) - LeetSync

+++ 152-maximum-product-subarray/maximum-product-subarray.cpp
@@ -13,6 +13,6 @@ public:
             ans=max(ans, max(prefix,suffix));
         }
 
-        return (int)ans;
+        return ans;
     }
 };
